requirements.txt
streamlit_elements==0.1.0
Pillow
langid
lingua-language-detector
codeswitch
transformers
torch
scikit-learn
matplotlib
plotly
fastdtw